Multiple panel support
To create matrices larger than 16*16, e.g. 4 panels of 8*32 each. In this update a panel is assumed to be 256 leds so nr of panels is calculated based on the total nr of leds. (e.g. 256 leds: 1 panel, 1024 leds => 4 panels).
Includes small update of Game of Life effect to support big matrices.