Support multiple panels v2.0
Improvements:
- if only 1 panel, algorithm is the same as original
- tested in wokwi
Note: in this version, overlayMin and Max of the Time & Macros settings screen is 'borrowed' to test the panels. In future update 3 variables should be added to Led Preferences screen (Multiple Panels checkbox and Horizontal and Vertical panels number). The first 3 lines of code of this XY function can then be removed